requirement.txt
streamlit
openai-whisper
